requirements.txt
openai>=1.0.0
google-generativeai>=0.3.0
openai-whisper
ffmpeg-python>=0.2.0
python-dotenv>=1.0.0
PyQt6
qdarkstyle
dotenv
